fix: made required changes
